+++ database/migrations/2025_04_24_141646_add_activity_id_to_quiz_contents_table.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Helpers\MigrationHelper;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
@@ -11,9 +12,14 @@ return new class extends Migration
      */
     public function up(): void
     {
-       Schema::table('quiz_contents', function (Blueprint $table) {
-        $table->foreignId('activity_id')->after('id')->constrained('activities')->onDelete('cascade');
-    });
+        Schema::table('quiz_contents', function (Blueprint $table) {
+            // Check if column already exists
+
+            if (!MigrationHelper::columnExists('quiz_contents', 'activity_id')) {
+
+                $table->foreignId('activity_id')->after('id')->constrained('activities')->onDelete('cascade');
+            }
+        });
     }
 
     /**
